// src/API.ts
export type CreateProductInput = {
  id?: string | null,
  image: string,
  images: Array< string >,
  brand: string,
  price: number,
  rating: number,
  numReview: number,
  isFeatured?: boolean | null,
  name: string,
  description: string,
  countInStock?: number | null,
  _version?: number | null,
  categoryProductsId?: string | null,
};

export type ModelProductConditionInput = {
  image?: ModelStringInput | null,
  images?: ModelStringInput | null,
  brand?: ModelStringInput | null,
  price?: ModelFloatInput | null,
  rating?: ModelFloatInput | null,
  numReview?: ModelIntInput | null,
  isFeatured?: ModelBooleanInput | null,
  name?: ModelStringInput | null,
  description?: ModelStringInput | null,
  countInStock?: ModelIntInput | null,
  and?: Array< ModelProductConditionInput | null > | null,
  or?: Array< ModelProductConditionInput | null > | null,
  not?: ModelProductConditionInput | null,
  _deleted?: ModelBooleanInput | null,
  createdAt?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
  categoryProductsId?: ModelIDInput | null,
};
// ...
export type Product = {
  __typename: "Product",
  id: string,
  image: string,
  images: Array< string >,
  brand: string,
  price: number,
  rating: number,
  numReview: number,
  isFeatured?: boolean | null,
  name: string,
  description: string,
  countInStock?: number | null,
  category?: Category | null,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
  categoryProductsId?: string | null,
};

export type Category = {
  __typename: "Category",
  id: string,
  name: string,
  src: string,
  products?: ModelProductConnection | null,
  createdAt: string,
  updatedAt: string,
  _version: number,
  _deleted?: boolean | null,
  _lastChangedAt: number,
};

export type ModelProductConnection = {
  __typename: "ModelProductConnection",
  items:  Array<Product | null >,
  nextToken?: string | null,
  startedAt?: number | null,
};

export type UpdateProductInput = {
  id: string,
  image?: string | null,
  images?: Array< string > | null,
  brand?: string | null,
  price?: number | null,
  rating?: number | null,
  numReview?: number | null,
  isFeatured?: boolean | null,
  name?: string | null,
  description?: string | null,
  countInStock?: number | null,
  _version?: number | null,
  categoryProductsId?: string | null,
};

export type DeleteProductInput = {
  id: string,
  _version?: number | null,
};

export type CreateCategoryInput = {
  id?: string | null,
  name: string,
  src: string,
  _version?: number | null,
};

export type ModelCategoryConditionInput = {
  name?: ModelStringInput | null,
  src?: ModelStringInput | null,
  and?: Array< ModelCategoryConditionInput | null > | null,
  or?: Array< ModelCategoryConditionInput | null > | null,
  not?: ModelCategoryConditionInput | null,
  _deleted?: ModelBooleanInput | null,
  createdAt?: ModelStringInput | null,
  updatedAt?: ModelStringInput | null,
};

export type UpdateCategoryInput = {
  id: string,
  name?: string | null,
  src?: string | null,
  _version?: number | null,
};
// ...
export type CreateProductMutationVariables = {
  input: CreateProductInput,
  condition?: ModelProductConditionInput | null,
};

export type CreateProductMutation = {
  createProduct?:  {
    __typename: "Product",
    id: string,
    image: string,
    images: Array< string >,
    brand: string,
    price: number,
    rating: number,
    numReview: number,
    isFeatured?: boolean | null,
    name: string,
    description: string,
    countInStock?: number | null,
    category?:  {
      __typename: "Category",
      id: string,
      name: string,
      src: string,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
    categoryProductsId?: string | null,
  } | null,
};

export type UpdateProductMutationVariables = {
  input: UpdateProductInput,
  condition?: ModelProductConditionInput | null,
};

export type UpdateProductMutation = {
  updateProduct?:  {
    __typename: "Product",
    id: string,
    image: string,
    images: Array< string >,
    brand: string,
    price: number,
    rating: number,
    numReview: number,
    isFeatured?: boolean | null,
    name: string,
    description: string,
    countInStock?: number | null,
    category?:  {
      __typename: "Category",
      id: string,
      name: string,
      src: string,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
    categoryProductsId?: string | null,
  } | null,
};

export type DeleteProductMutationVariables = {
  input: DeleteProductInput,
  condition?: ModelProductConditionInput | null,
};

export type DeleteProductMutation = {
  deleteProduct?:  {
    __typename: "Product",
    id: string,
    image: string,
    images: Array< string >,
    brand: string,
    price: number,
    rating: number,
    numReview: number,
    isFeatured?: boolean | null,
    name: string,
    description: string,
    countInStock?: number | null,
    category?:  {
      __typename: "Category",
      id: string,
      name: string,
      src: string,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
    categoryProductsId?: string | null,
  } | null,
};

export type CreateCategoryMutationVariables = {
  input: CreateCategoryInput,
  condition?: ModelCategoryConditionInput | null,
};

export type CreateCategoryMutation = {
  createCategory?:  {
    __typename: "Category",
    id: string,
    name: string,
    src: string,
    products?:  {
      __typename: "ModelProductConnection",
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
  } | null,
};

export type UpdateCategoryMutationVariables = {
  input: UpdateCategoryInput,
  condition?: ModelCategoryConditionInput | null,
};

export type UpdateCategoryMutation = {
  updateCategory?:  {
    __typename: "Category",
    id: string,
    name: string,
    src: string,
    products?:  {
      __typename: "ModelProductConnection",
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
  } | null,
};

export type DeleteCategoryMutationVariables = {
  input: DeleteCategoryInput,
  condition?: ModelCategoryConditionInput | null,
};

export type DeleteCategoryMutation = {
  deleteCategory?:  {
    __typename: "Category",
    id: string,
    name: string,
    src: string,
    products?:  {
      __typename: "ModelProductConnection",
      nextToken?: string | null,
      startedAt?: number | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
  } | null,
};

export type CreateFavoriteProductMutationVariables = {
  input: CreateFavoriteProductInput,
  condition?: ModelFavoriteProductConditionInput | null,
};

export type CreateFavoriteProductMutation = {
  createFavoriteProduct?:  {
    __typename: "FavoriteProduct",
    id: string,
    userSub: string,
    productID: string,
    product?:  {
      __typename: "Product",
      id: string,
      image: string,
      images: Array< string >,
      brand: string,
      price: number,
      rating: number,
      numReview: number,
      isFeatured?: boolean | null,
      name: string,
      description: string,
      countInStock?: number | null,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
      categoryProductsId?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
    favoriteProductProductId?: string | null,
  } | null,
};

export type UpdateFavoriteProductMutationVariables = {
  input: UpdateFavoriteProductInput,
  condition?: ModelFavoriteProductConditionInput | null,
};

export type UpdateFavoriteProductMutation = {
  updateFavoriteProduct?:  {
    __typename: "FavoriteProduct",
    id: string,
    userSub: string,
    productID: string,
    product?:  {
      __typename: "Product",
      id: string,
      image: string,
      images: Array< string >,
      brand: string,
      price: number,
      rating: number,
      numReview: number,
      isFeatured?: boolean | null,
      name: string,
      description: string,
      countInStock?: number | null,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
      categoryProductsId?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
    favoriteProductProductId?: string | null,
  } | null,
};

export type DeleteFavoriteProductMutationVariables = {
  input: DeleteFavoriteProductInput,
  condition?: ModelFavoriteProductConditionInput | null,
};

export type DeleteFavoriteProductMutation = {
  deleteFavoriteProduct?:  {
    __typename: "FavoriteProduct",
    id: string,
    userSub: string,
    productID: string,
    product?:  {
      __typename: "Product",
      id: string,
      image: string,
      images: Array< string >,
      brand: string,
      price: number,
      rating: number,
      numReview: number,
      isFeatured?: boolean | null,
      name: string,
      description: string,
      countInStock?: number | null,
      createdAt: string,
      updatedAt: string,
      _version: number,
      _deleted?: boolean | null,
      _lastChangedAt: number,
      categoryProductsId?: string | null,
    } | null,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
    favoriteProductProductId?: string | null,
  } | null,
};

export type CreateMessageMutationVariables = {
  input: CreateMessageInput,
  condition?: ModelMessageConditionInput | null,
};

export type CreateMessageMutation = {
  createMessage?:  {
    __typename: "Message",
    id: string,
    image: string,
    productName: string,
    sellerName: string,
    situation: string,
    date: string,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
  } | null,
};

export type UpdateMessageMutationVariables = {
  input: UpdateMessageInput,
  condition?: ModelMessageConditionInput | null,
};

export type UpdateMessageMutation = {
  updateMessage?:  {
    __typename: "Message",
    id: string,
    image: string,
    productName: string,
    sellerName: string,
    situation: string,
    date: string,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
  } | null,
};

export type DeleteMessageMutationVariables = {
  input: DeleteMessageInput,
  condition?: ModelMessageConditionInput | null,
};

export type DeleteMessageMutation = {
  deleteMessage?:  {
    __typename: "Message",
    id: string,
    image: string,
    productName: string,
    sellerName: string,
    situation: string,
    date: string,
    createdAt: string,
    updatedAt: string,
    _version: number,
    _deleted?: boolean | null,
    _lastChangedAt: number,
  } | null,
};